Add a new storage OTP application.
Currently it uses mnesia as a storage backend. mnesia can act as a
distributed database, and the initialisation code reflects this.
It checks the number of nodes, and based on result, it either sets up
the database(if it is the first node), or it adds itself as a node.

This is not heavily tested(or complete) at this point, and needs to be
made production-ready.

Notes:
mnesia chooses CP in the CAP model, so sync can take time in case of netsplit.
Merges may need to be manual.
This is probably okay, as long as we just keep only config in mnesia.
mnesia allows us to have proxy42 as a completely embedded program.

But other options, like riak also need to be tested.