embed component applications into proxy42
OTP applications can be stopped independent of each other, unless they are
permanent. This allows proxy42_core to exit or be stopped for whatever reason
while proxy42 still shows up as running. While it is easy to track such issues
now, with many more components it my be a bit unwieldy. Since the components are
such that they are not useful outside proxy42, we choose to include them.

In addition, stopping proxy42 should stop all component applications, which
wasn't the case. Including solves this issue as well. This could also have been
done by stopping the components when proxy42 is stopped, but the above issue
will still remain. Monitoring other applications sanely requires using
undocumented apis, which we are trying to avoid.

Though including is generally frowned upon we believe this is an appropriate
use case.

An included application is loaded by application controller but not started.
The lifecycle of an included application is the responsibility of including
application. While not required, it is a good idea to attach included
applications to the supervision tree of including application.