[core] refactor auth, ratelimit plugins to accept ReqMeta
The state of proxy42_router contains request metadata that is expected to be
updated as the request progresses through various stages of identifications,
checks and balances. All plugins will receive this metadata and can make
arbitrarily complex decisions based on information that is known about the
request until that point in the processing pipeline.

Co-authored-by: Saurabh Kumar <github@sa1.me>