rename domain_group to api throughout
Domain group was a term coined by heroku, and it suits their use cases
fine. However, we intend to have a clean logical separation between a
domain and an api. The term domain group becomes confusing. Since it
was used to hold what would eventually be api, we're renaming it to
api

Co-Authored-By: Saurabh Kumar <github@sa1.me>