add stubs for multi level settings
We plan to support settings at domain, api, request, developer and
application user levels. As the request flows through various stages of
its execution, we update the request context with all available settings
until that point.

Some of the advantages:
This allows defaults to be set at domain or global levels.
This allows restricting allowed strategies in lower levels.
This allows overriding authentications or rate limits based on request
method.

Co-Authored-By: Saurabh Kumar <github@sa1.me>